src/requirements.txt
wheel
flask
flask-socketio
RPi.GPIO
numpy
Pillow
picamera
ilock
PyYAML